reviewer/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "reviewer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "restate-sdk>=0.15.0",
    "pydantic-ai[openai]>=0.2.0",
    "pydantic>=2.0",
    "hypercorn>=0.17.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["reviewer*"]
